extracted settings card and text button and used max width with main layout div
